+++ Gui.py
@@ -4,7 +4,8 @@ from PyQt5.QtCore import Qt
 from speechtotext import listen
 from texttospeech import speak_this
 from groot import input_taking
-from utils.newsLibrary import news_func
+from newsWindow import newsBox
+from asset.modules.newsLib.newsLibrary import news_func
 import urllib.request
 
 
@@ -17,26 +18,18 @@ class Groot_Ui(QWidget):
         self.mainLayout = QVBoxLayout()
         self.upperLayout = inpOutWidgetLayout()
         self.mainLayout.addLayout(self.upperLayout.inpOutLayout)
-        #self.mainLayout.addStretch(1)
+        self.mainLayout.addStretch(1)
         
-        self.lowerLayout = TabLayout()
-        self.mainLayout.addWidget(self.lowerLayout.customTab)
-
         self.setLayout(self.mainLayout)
         self.setWindowTitle('GROOT')
         self.setWindowIcon(QIcon('asset/img/groot.png'))
-        self.setGeometry(150,150,800,500)
+        self.setGeometry(150,150,800,100)
         self.show()
 
         #Other Functions:
         self.upperLayout.speaker.clicked.connect(self.listen_reply)
-        # self.upperLayout.refresh.clicked.connect(self.createPopup)
+        self.upperLayout.newsButton.clicked.connect(self.createNewsLayout)
         self.upperLayout.inputEdit.returnPressed.connect(self.callGroot)
-
-
-    # def createPopup(self):
-    #     popup = Popup()
-    #     popup.show()
 
     def callGroot(self):
         typed = str(self.upperLayout.inputEdit.text())
@@ -52,6 +45,11 @@ class Groot_Ui(QWidget):
         if e.key() == Qt.Key_Escape:                
             self.close()
 
+    def createNewsLayout(self):
+
+        self.newsWindow = newsBox()
+        self.newsWindow.show()     
+    
 
     #Function to display warning message for quitting
     def closeEvent(self, event):
@@ -85,9 +83,9 @@ class inpOutWidgetLayout(QVBoxLayout):
         #Text Field for Output
         self.outputEdit = QLineEdit()
         self.outputEdit.setMinimumSize(self.sizeHint())
-        #Refresh Button
-        self.refresh  = QPushButton('Refresh')
-        self.refresh.setMinimumSize(self.sizeHint())
+        #newsButton Button
+        self.newsButton  = QPushButton('News')
+        self.newsButton.setMinimumSize(self.sizeHint())
         #Spacing between the widgets
         self.inpOutLayout.setSpacing(20)
         #Positioning Inner Widgets
@@ -96,64 +94,10 @@ class inpOutWidgetLayout(QVBoxLayout):
         self.inpOutLayout.addWidget(self.speaker,1,2)
         self.inpOutLayout.addWidget(self.outputLabel, 2, 0)
         self.inpOutLayout.addWidget(self.outputEdit, 2, 1)
-        self.inpOutLayout.addWidget(self.refresh,2,2)
+        self.inpOutLayout.addWidget(self.newsButton,2,2)
 
         #Button is needed in parallel.
         
         
 
-class TabLayout(QTabWidget):
-    def __init__(self,parent=None):
-        super(TabLayout,self).__init__()
-        self.customTab = QTabWidget()
-        self.news = QScrollArea()
-        #self.news.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.scroll_widget = QWidget()
-        self.scroll_layout = QVBoxLayout(self.scroll_widget)
-        self.scroll_layout.setSpacing(20.0)
-        articles = news_func()
 
-        for article in articles:
-            self.myLabel = Tiles(parent="None",description = article['description'][0:100],title=article['title'],url=article['url'],urlImage=article['urlToImage'])
-            self.scroll_layout.addLayout(self.myLabel.tileLayout)
-
-        self.news.setWidget(self.scroll_widget)
-
-        self.wiki = QWidget()
-        
-        self.customTab.addTab(self.news,"News")
-        self.customTab.addTab(self.wiki,"Add More Feature")
-
-class Tiles(QWidget):
-    def __init__(self,parent="None",title="None",description="None",url="None",urlImage="None"):
-        super(Tiles,self).__init__()
-        #Heading Widget
-        self.heading = QLabel('<b>%s</b>'%title)
-        #SubHeading Widget with link to open in browser
-        self.subHeading = QLabel('{}<a href="{}">...more</a>'.format(description,url))
-        self.subHeading.setOpenExternalLinks(True)
-        #Image Widget with article
-        data = urllib.request.urlopen(urlImage).read()
-        self.image = QImage()
-        self.image.loadFromData(data)
-        self.imageLabel = QLabel("image")
-        self.imageLabel.setPixmap(QPixmap(self.image).scaled(64,64,Qt.KeepAspectRatio))
-        self.tileLayout = QGridLayout()
-        self.tileLayout.addWidget(self.heading,1,0)
-        self.tileLayout.addWidget(self.imageLabel,1,1)
-        self.tileLayout.addWidget(self.subHeading,2,0)
-
-# class Popup(QWidget):
-#     def __init__(self, parent=None):
-#         super(Popup,self).__init__()
-#         self.PopupLayout = QGridLayout()
-#         self.title("Hi I am Groot Why Don't you add a Keyword to search in news?")
-#         self.keywordField = QLineEdit()
-#         self.keywordField.returnPressed.connect(closemyself)
-
-#         def closemyself():
-#             typed = str(self.keywordField.text())
-
-#         self.setGeometry(150,150,800,500)
-#         print("I am created")
-#         self.show()
